Restructured API project
