+++ API/Tenant/Requests/RegisterTenant.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace API.Requests.Tenant
+﻿namespace API.Tenant.Requests
 {
     public class Register
     {
